+++ packages/CHANGELOG.md
@@ -3,6 +3,11 @@
 
 ### Next Release
 
+**`rev-models`**
+
+* Added support for fields with property getters. By default, RevJS fields
+  with a getter function are set to not be stored (`Field.options.stored == false`)
+
 **`rev-api-client`**
 
 * Add support for `orderBy` on `read()`